Update luftdaten.markdown
The setup process is different for sensor owners...
I'm not an expert... I guess the motherboard sensor UUID (Example esp8266-1XX9XXX7) is mapped to a first sensorId (Example 39XXX) which is shown internal in the user account as the main ID marked with a #.

This ID returns only some of the entities available at Luftdaten.info / sensor.community / the local sensor.
Better results (more entities) can be found with the *Sensor ID* found in the *Data* Part...

If the local Sensor webpage shows data not shown on sensor.community - (Example the temperature / humidity is empty there but available localy) Check the *Settings* if the correct Sensor is selected. (Example DHT22, BME280 ...)

The public available sensor data on the map might be different to this data. My guess... If the sensor owner does not share all the sensor data - the id is different again.